Properly close open files and check for missing files.
Also changed size_t no to uint32_t.

+++ executor.c
@@ -2,12 +2,12 @@
 #include "joblist.h"
 #include "executor.h"
 
-static int job_counter = 0;
+static uint32_t job_counter = 0;
 
 /* Returns the next available job number.
  * The first job nubmer is always 1.
  */
-int get_next_job_number()
+uint32_t get_next_job_number()
 {
 	return ++job_counter;
 }
@@ -27,7 +27,7 @@ int start_job(job_t *job)
 		parallel_char = '*';
 
 	printf("Job %d%c: <%s> ",job->job_n,parallel_char,job->prog_name);
-	for(size_t i = 0;i < job->size_params;++i)
+	for(uint32_t i = 0;i < job->size_params;++i)
 	{
 		printf("[%s] ",job->params[i]);
 	}
